+++ src/bsoncxx/include/bsoncxx/v_noabi/bsoncxx/builder/list.hpp
@@ -30,16 +30,6 @@ namespace bsoncxx {
 namespace v_noabi {
 namespace builder {
 
-using namespace ::bsoncxx::v_noabi::types;  // Deprecated. Deliberately undocumented.
-
-}  // namespace builder
-}  // namespace v_noabi
-}  // namespace bsoncxx
-
-namespace bsoncxx {
-namespace v_noabi {
-namespace builder {
-
 ///
 /// A JSON-like builder for creating documents and arrays.
 ///
@@ -94,7 +84,7 @@ class list {
     /// @see
     /// - @ref bsoncxx::v_noabi::types::bson_value::view.
     ///
-    operator bson_value::view() {
+    operator types::bson_value::view() {
         return view();
     }
 
@@ -104,12 +94,12 @@ class list {
     /// @see
     /// - @ref bsoncxx::v_noabi::types::bson_value::view.
     ///
-    bson_value::view view() {
+    types::bson_value::view view() {
         return val.view();
     }
 
    private:
-    bson_value::value val;
+    types::bson_value::value val;
 
     friend ::bsoncxx::v_noabi::builder::document;
     friend ::bsoncxx::v_noabi::builder::array;
@@ -141,12 +131,12 @@ class list {
                 _core.key_owned(std::string((begin(init) + i)->val.view().get_string().value));
                 _core.append((begin(init) + i + 1)->val);
             }
-            val = bson_value::value(_core.extract_document());
+            val = types::bson_value::value(_core.extract_document());
         } else if (type_deduction || is_array) {
             core _core{true};
             for (auto&& ele : init)
                 _core.append(ele.val);
-            val = bson_value::value(_core.extract_array());
+            val = types::bson_value::value(_core.extract_array());
         } else {
             throw bsoncxx::v_noabi::exception{error_code::k_unmatched_key_in_builder,
                                               err_msg.str()};
@@ -205,14 +195,6 @@ class array : public list {
 };
 }  // namespace builder
 }  // namespace v_noabi
-}  // namespace bsoncxx
-
-namespace bsoncxx {
-namespace builder {
-
-using namespace ::bsoncxx::v_noabi::types;  // Deprecated. Deliberately undocumented.
-
-}  // namespace builder
 }  // namespace bsoncxx
 
 #include <bsoncxx/config/postlude.hpp>